Fix: LocationInfo POJO had incorrect JsonKey mapping of _id. Fix: GeoPosition POJO was mapping latitude and longitude values as long instead of double. Add: Writing to CSV.

+++ src/main/java/com/cassarmanghi/experiment/goeurodevtestoct15/pojos/GeoPosition.java
@@ -13,15 +13,15 @@ public class GeoPosition {
   }
 
   @SerializedName(JsonKeys.LATITUDE)
-  private long latitude;
+  private double latitude;
   @SerializedName(JsonKeys.LONGITUDE)
-  private long longitude;
+  private double longitude;
 
-  public long getLatitude() {
+  public double getLatitude() {
     return latitude;
   }
 
-  public long getLongitude() {
+  public double getLongitude() {
     return longitude;
   }
 }
